Include patch plugin for open url
Some of the menu items in the default Help menu use an older `open_url`
command to open the appropriate URL. Under Linux this does not properly
select the default browser.

This patch plugin replaces the built in command with one that functions
identically but which uses the `webbrowser` module, which allows it to
work as expected.

For more information:
    https://forum.sublimetext.com/t/how-does-the-open-url-command-select-the-browser-to-use/21590