Include some initial files for Common Questions
This sets up the repository for files in the Common Questions series,
which at the moment is only the demonstration custom build target for
the video that will be dropping imminently.

It may be worth it to go back over some of the previous videos and see
if any of the custom builds done thus far may be worth presenting in
this manner as well, though I think they're all single purpose and
demonstrative enough that this is not strictly needed.