Include the plugin for Common Questions 10
The subject of this video is creating simple custom build target that
takes some initial action (such as expanding a custom variable) and
then uses the `exec` command to perform the build as per normal
afterwards.

This is the final sample code from the video, in which there is a
confirmation check made to see if the build should trigger. The command
also allows for a custom build key named `caption` that allows you to
change the prompt in the dialog box, which can contain any of the
standard build system variables.