Include the plugin for Common Questions 11
The subject of this video is canceling a running build by using the
`cancel` key in the `sublime-build` file.

This is a very slight extension to the plugin provided in CQ10; here we
have a simple short circuit that allows the custom target command to
still prompt you to start the build but not prompt you to cancel it.

With simple tweaks, this could also be extended to asking you to confirm
a cancel instead, if desired.