Include the plugin for Common Questions 12
The subject of this video is using a more advanced technique to create
a custom build target that can be used to take actions when the build
is finished.

This is a modified version of the plugin provided in CQ10; here we
leverage the internal `exec` command by subclassing it directly so that
we can use existing code to determine when (and how) the build finishes.