Include the plugin for Common Questions 13
This video covers some real world examples and use cases for customized
versions of the internal `exec` command for use in build systems.

The first example shows how to temporarily suppress the build output
panel display on build execution, which is handy for executing external
programs where you don't care about the build output.

The second and third examples show how to capture the selected text and
provide it to the build system as a command line argument or as a
temporary file for the external command to use.